+++ modules/duckdns.nix
@@ -1,9 +1,12 @@
 {
   nixosModule = { pkgs, ... }: {
+    age.secrets.duckdns.file = ../secrets/duckdns.age;
+
     systemd.services.duckdns = {
       description = "Update DuckDNS";
       serviceConfig = {
-        ExecStart = "${pkgs.curl}/bin/curl 'https://www.duckdns.org/update?domains=enzime&token=d725755a-58af-4ca0-a28a-56a844640289&ip='";
+        EnvironmentFile = "/run/agenix/duckdns";
+        ExecStart = "${pkgs.curl}/bin/curl 'https://www.duckdns.org/update?domains=\${SUBDOMAIN}&token=\${TOKEN}&ip='";
       };
     };