+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Fixed `vim` file encoding breaking when using `git commit` with `age` encrypted files
+- Changed `duckdns` module to store `subdomain` and `token` as secrets
+- Use `agenix` for secrets in my dotfiles
 - Removed `delta`
 - Removed unnecessary `networking.hostName` in `phi/configuration.nix`
 - Added `tmux`